Make the BootKeyboard support optional
Introducing two macros: `WITH_BOOTKEYBOARD` and `WITH_BOOTKEYBOARD_PROTOCOL`.
Both must be followed by a block of code. The latter is to be used whenever we
want to use the `BootKeyboard` fallback mechanism; the former whenever we want
to call a method of `BootKeyboard`, regardless of the HID protocol set.

These macros expand differently, depending on whether
`KALEIDOSCOPE_HIDADAPTOR_DISABLE_KEYBOARD_BOOT_PROTOCOL` is set or not (unset
being the default). When set, the two macros above will expand to an `if (0)`
block, allowing the compiler to discard the code alltogether. When unset, they
expand differently, as explained in the comments.

These changes allow one to easily disable the `BootKeyboard` mechanism, by
setting a compiler flag, either via the IDE, or via the `LOCAL_CFLAGS` variable
used by `kaleidoscope-builder`.

Signed-off-by: Gergely Nagy <algernon@keyboard.io>