Don't use double-negatives for the knob checks
Instead of using `#if !FOO_DISABLE_BAR`, use `#if FOO_ENABLE_BAR`, and set the
enable flag in the main header, based on the disable knob. Do this to make the
ifs clearer, while still having an enabled-by-default, without having to
explicitly set the flag somewhere (be that in local CFLAGS or `boards.txt`).

This - I believe - gets us the best of both worlds.

Signed-off-by: Gergely Nagy <algernon@keyboard.io>