A fix for rollover from a mod-flagged key
When rollover occurs from a `Key` with a modifier flag to one without, the modifier flag
should not be applied to the new keypress. This change makes that possible by adding a
mask for mod flags, and only allowing the flags from the latest keypress to be added to
the report, with the exception of "pure" modifiers (anything with a modifier `keyCode`
value still gets its mod flags added, regardless).