A fix for rollover between keys with and without attached modifiers
When rollover occurs from a `Key` with a modifier flag to one without, the modifier flag
should not be applied to the new keypress. This change makes that possible by adding a
mask for mod flags, and only allowing the flags from the latest keypress to be added to
the report, with the exception of "pure" modifiers (anything with a modifier `keyCode`
value still gets its mod flags added, regardless).

If a combo-modifier key (a modifier that also has mod flags) toggles on in the same cycle
as a key without those mod flags, and the other key is processed after the combo-modifier,
the flags from the combo would be excluded from the report. To fix this, we call
`pressKey()` instead of setting `current_mod_flags`.

(@gedankenexperimenter is responsible for the design and implementation
of this feature, though @obra made changes to the documentation, API and
naming.)

Fixes keyboardio/Model01-Firmware#23
Fixes keyboardio/Model01-Firmware#61
Fixes keyboardio/Kaleidoscope#293