Fixed one simultaneous-press issue
If a combo-modifier key (a modifier that also has mod flags) toggles on in the same cycle
as a key without those mod flags, and the other key is processed after the combo-modifier,
the flags from the combo would be excluded from the report. To fix this, we call
`pressKey()` instead of just setting `current_mod_flags`. This is also probably more
understandable code.