Combine mod flags when two keys toggle on in the same cycle
This isn't really correct, but it's at least not regressive (there is no new incorrect
behaviour; at worst it's the same as before the rollover fix). There are solutions that
are more correct (either sending an extra report or delaying a keypress), but those are
much more complex, and I don't think they're worth the cost.