Corrected removal of modifier bits from modifier_flag_mask
Instead of using XOR, to remove only the bits set in `released_key.flags` when a key is
released, we need to flip the bits in `released_key.flags` and do a bitwise AND. With the
XOR, we can remove other flags that are set.

In practice, this is quite unlikely to ever come up, but the XOR is nevertheless incorrect.

Signed-off-by: Michael Richters <gedankenexperimenter@gmail.com>