Fix mod-flag rollover problem
This change reverts ff727ee, because we actually need the extra report for other
reasons. That change only sent the extra report without the newest keycode if
that keycode was also in the previous report, but we also need that extra report
if we've changed a modifier bit. If we send a single report which contains both
a keycode bit toggling on and a modifier toggling off, some host systems will
process the keypress first, causing the modifier to be applied to the initial
character generated.

Fixes #18.

Signed-off-by: Michael Richters <gedankenexperimenter@gmail.com>